Serialize all Nitrokey tests
Synchronization when accessing the Nitrokey has been a long-standing
problem, as individual requests are not synchronized and there is
potential for screwing up some internal device state unless they
serialize everything correctly on their end (which is questionable).
However, even that would not be enough: Rust unit tests run concurrently
by default and individual requests made by one test could interleave
with those of another, causing spurious test failures or worse.

There exist a variety of approaches to those problems. Basic per-device
synchronization in libnitrokey would be a proper starting point but does
not get us all the way to higher-level synchronized primitives.
Transaction or session support would be another one. With sessions we
could isolate "virtual" connections to the device which would be pretty
much exactly what we need.
None of this is on anybody's road map, though, because most applications
accessing the device are single threaded anyway (or at least requests
are issues by a single thread).
To accommodate for the case of interfering tests, this change implements
a solution that is entirely on the testing side. Since we already wrap
all tests connecting to the device, it is natural to guard device
accesses with a lock at this level. This change implements such a
scheme.